requirements.txt
jinja2
google-genai
python-dotenv
tqdm
nltk
ruff
pre-commit
